Initial Code Commit
Added the two main solutions auto-remind.sln and arum.tests.sln for the
core application and tests respectively.

Added the arum.core project that will house the core utilities and
objects to be used in the application. These still need some work  to
ensure they are up to the task (pun intended)

A note, I put in the EnumUtil class in the aurm.core dll, when I think
it really should go in the orth.core dll. I will have to add them to
that later and republish to NuGet.

Added arum.tests.sln and arum.core.tests.csproj to do unit tests on the
core dll. XUnit is the accepted standard for the project.