Added Tasker, a Task Watcher
Added Tasker, a class that implements ITaskMonitor that takes a
List<AurmTask> as its constructor and will "monitor" the task collection
on a 1 minute interval. If any of the AurmTasks has all of its
conditions "IsMet", then it will raise the TaskUpdated event.
This leads me to imagine the condition where the task is completed, but
still in the list. It will refire every minute. We also need a
LastFiredTime property in the AurmTask  that keeps track of the last
time it was fired, and only check the condition for ones that have not
fired today