Added Notification System
This is in the need of the most work.
Added a dll, arum.notifier which has two classes TwilioNotifier, which
will send SMSs to the Recipients and a SMSREcipient class.
These need further implementation, but I have yet to get started
learning the Twilio API.

At the very least, I need to add unit tests for the SMSRecipient's
constructor as it does Regex validation on the phone number passed