Give the user an opportunity to not notify
Now when the task event is raise, we MessageBox.Show and inform the user
that they might want to send a notification before we send it.

Also now flagging the LastTriggeredTime of the task before raising the
TaskUpdated event.

Added a quick note on the Json Net serialization to make sure we
deserialize properly