Final update to base functionality
Pulled out the phone number verifier to its own utility function into
the aurm.core.utilities namespace since it is used in multiple places.
Added unit tests for it as well.

In the TwilioNotifier, now getting the values for the Authentication
Key, Token and Phone Number from the App.config's appSettings.

This successfully satisfies all the original requirements I originally
set out to satisfy. Need to update the readme and consider rewriting
this in electron so it is cross platform :)